Define sourceId when creating quick tools from Okular settings
Only the quick annotation tools created using the "bookmark" action were correctly assigned the sourceId. Now also those created in the Okular settings page have this attribute correctly defined.

Without this field it was not possible to identify which built-in annotation tool action to check when the corresponding quick tool action is triggered by the user.

BUG: 424810
FIXED-IN: 1.11.0