Improve date handling for migration test
The dates in the test xml are in ISO format but without an explicit time zone.
This makes the documenttest fail due to differences in the literal string and
the UTC date.
Wrap the literal string in a date conversion to treat it as as a date.
Add the Z timezone designator for post migration due to differences in how
the date it treated in Okular depending on where it's read.