Fix annotation modificationDate to match okular timestamp
Restore original okl_ann modificationDate after addAnnotation
during which a new native pdfAnnot is created with current time
as timestamp.

This is not usually a problem when both create times are nearly
instanteous but can be problematic during annotation migration
when the okular modificationDate may be long before the just
created native Poppler annotation.