Add "Distraction-free Mode" under the "View" menu
When triggered, this mode will temporarily hide sidebar, menubar, main
toolbar and annotation toolbar but shows the pagebar. This allows for
more focused reading without having to manually hide all the different
toolbars by using the shortcuts one by one.

This is an alternative view to the "Presentation" view or "Fullscreen
Mode" which hide the taskbar of the window system.

Additional features

* Restores GUI to the state prior to Distraction-free Mode activation during next
  launch of Okular, if Distraction-free Mode is activated while closing a Shell.

* Distraction-free Mode is applied on Shell basis and is synchronized between tabs,
  if Tab mode is activated.

* When closing a shell in Distraction-free Mode with multiple tabs open, GUI
  state of the active tab is remembered for the next launch.

FEATURE: 325190
GUI: