PresentationWidget: Force rerendering if screen dpr changed
Consider a situation with two screens A and B, and A having
a higher device pixel ratio than B.  Place the main Okular
widget on screen A and set up presentation mode to open on
screen B.  In this situation, the presentation mode window
will show (partially) garbage, because the PresentationWidget
is first created on screen A and then moved to screen B.
The rendering still happens on screen A, which has the wrong
device pixel ratio.

The fix implemented here is to make PresentationWidget::paintEvent
check whether the device pixel ratio of the rendered page
is correct for the current screen, and to force a rerendering
if it is not.