Allow to print pdf doc directly into a QPrinter
This is the 'official' Qt way of printing. It avoids the conversion
of pdf files to postscript that we do currently, and may fix
a number of printer settings bugs.

There is a big caveat to this approach, however: Rendering a pdf doc
into a QPrinter object only works with the poppler 'QPainter' backend.
While that backend works okay on many documents, it is still lacking
important features, and you may see rendering problems. For the
most part this is not a fundamental problem. Somebody just needs
to sit down and write the missing parts.

For this reason, CUPS printing remains the default, and Qt printing
is simply a new option. Additionally, I tagged this option
as 'experimental' in the GUI.