Reset form text fields to last committed value when commit is rejected
When a commit is rejected by the keystroke commit script the form content should be reset to its last accepted/committed value

To do this any user change is first written to the form as a pending value.
When the commit is accepted the pending value becomes the new value.
When the commit is rejected the pending value is discarded.