Fixed Page Scroll Using Hires Scroll Wheel
This change fixes the issue of page scroll in non-continuous mode using
high resolution scroll wheel of a mouse. In the original code, the page
scroll is hardly triggered due to small delta values reported by the
mouse wheel. The modified code added a handle to accumulate the small
delta values to make sure page scroll will be triggered.

BUG: #443404