Fix annotation resizing
Current implementation of AnnotationPrivate::adjust changes deltas so that annotation's
bounding rectangle remains in [0,1] interval. This leads to error because deltas are
already stored in AdjustAnnotationCommand, meaning that we now have two different pairs of
values for deltas (one apllied to the annotation and one stored in AdjustAnnotationCommand).

The error can be exposed by moving right edge of annotation over the left and than using undo.

For that reason, I moved checks for deltas in MouseAnnotation::performCommand (where deltas
are first used).