No need to lookup images that doesn't exist
if there is no suffix, the image reader in Qt cycles thru all of the
possible values that it might be able to load to see if such a thing
exists relative to current workdir. This does give an extra check of
file in the good case, but avoids number-of-supported-image-formats
checks in the bad case.

While this is not a problem for most users, on certain so-called
enterprise systems with encrypted file systems and virus scanners and so
on, it makes a slight difference.