Rework UX for adding digital signature
Currently when adding a signature the user is prompted to specify a
rectangle, then chose a certificate. The problem here is that the
user doesn't know what size is appropriate, so it takes some trial and
error to get something good.

To address this change the flow a bit. First ask which certificate and
other parameters should be used. With that information we can show a
"preview" for the signature, helping the user pick a good size.

This makes the warning about a too small signature rect redundant since
the user sees that the signature won't fit. It was only an approximation
anyway.

The preview is only an approximation of the final appearance though.
Getting the actual appearance from Poppler would involve some larger
work.

CCBUG: 443403