Rework UX for adding digital signature
Currently when adding a signature the user is prompted to specify a
rectangle, then chose a certificate. The problem here is that the
user doesn't know what size is appropriate, so it takes some trial and
error to get something good.

To address this change the flow a bit. First ask which certificate and
other parameters should be used. With that information we create a
signature annotation object and add that to the page. The user can
then move and resize it like any annotation util the size and placement
is right.

Once the user is happy with the placement they click on a Finish Signing
button that then prompts for where to save the signed document and signs
it.

This makes the warning about a too small signature rect redundant since
the user sees that the signature won't fit. It was only an approximation
anyway.

CCBUG: 443403