cppcheck: simplify suppressions
the d_func hierarchy is quite common and should just be generalized.
That way, the suppressions might be able to be moved centrally

Also function argument pointers constness is something one should really
consider if it is worth fixing (it often isn't) due to issues with
binary compatibility or if the base or interface class might define it
without const pointers.